o1vm/riscv32im: simplify mod_signed

+++ o1vm/src/interpreters/riscv32im/witness.rs
@@ -521,9 +521,9 @@ impl<Fp: Field> InterpreterEnv for Env<Fp> {
         y: &Self::Variable,
         position: Self::Position,
     ) -> Self::Variable {
-        let x: u32 = (*x).try_into().unwrap();
-        let y: u32 = (*y).try_into().unwrap();
-        let res = ((x as i32) % (y as i32)) as u32;
+        let x: i32 = (*x).try_into().unwrap();
+        let y: i32 = (*y).try_into().unwrap();
+        let res = (x % y) as u32;
         let res = res as u64;
         self.write_column(position, res);
         res
